Trigger group creation on ust app registration
Create a pipe for each app and setup a trigger group associated with that
pipe. Transfer the write side to the app, and transfer the read side to
the notification thread as an application event source
(*_ADD_APPLICATION)

TODO: what if the app does not support trigger group?
TODO: handle pipe management on unregistration etc.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I3e4aab84e3270ddef1f50f72f946a4d80b3f36e0