Move kernel notification fd ownership to main thread
The notification thread teardown and cleanup is delayed compare to the
kernel tracer lifetime and modules. This might prevent the module to
unload. The notification fd is now initialized at the same time as the
group fd and closed on kernel tracer teardown.

How does that work we there is a late kernel initialization??? Same goes
for the kernel_consumer_monitor_pipe....

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I3520b7ecf781e4a945fd8a8f129d8d5a92f15b0e