CONTRIBUTING.md: clarify the guidelines for commit messages
In order to stream-line the code review process, I am adding a more
detailed explanation of the desired commit message format.

Of note, the commit title format `Fix: sub-system`, used informally
for a couple of months, is adopted for bug fixes. A template of the
sections expected in the commit message body of those patches is also
included.

More general guidelines are also added for feature contributions.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: Id0097949ca984e88367e8e78cb6111adb30c5b99