Generate capture bytecode on trigger reception
FIXME/TODO:
======
all this is unnecessary if we enforce the set notion at the condition
event rule level. This code is mostly a results of the notion of
multiple notify action and that multiple capture were possible across a
group of action.

If a condition event-rule can mention that the capture expression
already exist on append mos of this goes away. Only the bytecode
generation is left. The bytcode would set at the `struct
lttng_capture_descriptor` level on trigger reception.

Unless we plan on having multiple conditions for a single trigger.
======

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Ibf86096bd17be078cfc5e8bcd82347797389d448
Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479