event-expr: implement ref counting
Move the event-expr.c to libcommon to consolidate implementation related to
event-expr and to prevent a circular linking on liblttngctl to access
the lttng_event_expr_destroy/get/put functions. Nothing prevents us
from having another implementation file on the src/common side but this
is getting tedious for nothing.

The refcounting will be used during the creation of a unique set of
capture field for the tracer enabler.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I0dd2d258e62183fa88f57de4b54988584c609f6b
Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479